se creo el archivo multa.py y se modularizó la función para calcular el valor de la multa
